+++ EleksInternshipProj.Server/EleksInternshipProj.Domain/Abstractions/INotificationRepository.cs
@@ -7,5 +7,7 @@ namespace EleksInternshipProj.Domain.Abstractions
         Task<IEnumerable<Notification>> GetSpaceNotificationsAsync(long spaceId);
 
         Task AddNotificationAsync(Notification notification);
+
+        Task<bool> ExitstForRelatedAsync(string relatedType, long relatedId);
     }
 }
